refactor : DetailModal에서 StarRate클래스 분리
