+++ src/components/modal/DetailModal.js
@@ -1,10 +1,9 @@
-import MovieRatingStorage from '../../utils/storage/MovieRatingStorage.js';
+
+import StarRating from './StarRating.js';
 
 class DetailModal {
   constructor(movie) {
     this.movie = movie;
-    this.ratingStorage = new MovieRatingStorage();
-    this.userRating = this.ratingStorage.getRating(this.movie.id) || 0;
   }
 
   render() {
@@ -40,21 +39,8 @@ class DetailModal {
             <hr />
             <p class="my-rate">
               <span>내 별점</span>
-              <div class="star-rating">
-              <div class="star-rating-item">
-                <span class="empty-star" data-rating="1"><img src="./images/star_empty.png" class="star" /></span>
-                <span class="empty-star" data-rating="2"><img src="./images/star_empty.png" class="star" /></span>
-                <span class="empty-star" data-rating="3"><img src="./images/star_empty.png" class="star" /></span>
-                <span class="empty-star" data-rating="4"><img src="./images/star_empty.png" class="star" /></span>
-                <span class="empty-star" data-rating="5"><img src="./images/star_empty.png" class="star" /></span>
-                </div>
-                <div class = "detail-rating">
-                <span class="rating-text">이 작품 어땠나요?</span>
-                <span class="rating-score">(?/10)</span>
-                </div>
-              </div>
+              <div class="star-rating-container"></div>
             </p>
-
 
             <hr />
             <div class="movie-overview">
@@ -68,68 +54,15 @@ class DetailModal {
         </div>
       </div>
     `;
-    this.setupStarRating(modalBackground);
+
+    // StarRating 컴포넌트 추가
+    const starRatingContainer = modalBackground.querySelector('.star-rating-container');
+    const starRating = new StarRating(this.movie);
+    starRatingContainer.appendChild(starRating.render());
 
     return modalBackground;
   }
 
-  setupStarRating(modalElement) {
-    const stars = modalElement.querySelectorAll('.empty-star');
-    const ratingText = modalElement.querySelector('.rating-text');
-    const ratingScore = modalElement.querySelector('.rating-score');
-
-    const ratingTexts = {
-      1: '최악이에요',
-      2: '별로예요',
-      3: '보통이에요',
-      4: '재미있어요',
-      5: '명작이에요',
-    };
-
-    const ratingScores = {
-      1: '(2/10)',
-      2: '(4/10)',
-      3: '(6/10)',
-      4: '(8/10)',
-      5: '(10/10)',
-    };
-
-    if (this.userRating > 0) {
-      stars.forEach((s, index) => {
-        if (index < this.userRating) {
-          s.innerHTML = '<img src="./images/star_filled.png" class="star" />';
-          s.classList.remove('empty-star');
-          s.classList.add('filled-star');
-        }
-      });
-      ratingText.textContent = ratingTexts[this.userRating];
-      ratingScore.textContent = ratingScores[this.userRating];
-    }
-
-    stars.forEach(star => {
-      star.addEventListener('click', () => {
-        const rating = parseInt(star.dataset.rating);
-        this.userRating = rating;
-        this.ratingStorage.saveRating(this.movie.id, rating);
-
-        stars.forEach((s, index) => {
-          if (index < rating) {
-            s.innerHTML = '<img src="./images/star_filled.png" class="star" />';
-            s.classList.remove('empty-star');
-            s.classList.add('filled-star');
-          } else {
-            s.innerHTML = '<img src="./images/star_empty.png" class="star" />';
-            s.classList.remove('filled-star');
-            s.classList.add('empty-star');
-          }
-        });
-
-        // 텍스트와 점수 업데이트
-        ratingText.textContent = ratingTexts[rating];
-        ratingScore.textContent = ratingScores[rating];
-      });
-    });
-  }
 
   addDetailModal(movie) {
     const modal = new DetailModal(movie);
